(Рефактор): Изменение структуры на много-модульный (Чистка): Перенесение функции загрузки файла через КлассЛоудер в ютил
