+++ modules/importer/src/main/kotlin/org/politecon/util/MiscUtil.kt
@@ -1,0 +1,3 @@
+package org.politecon.util
+
+fun getResourceAsText(path: String): String? = object {}.javaClass.getResource(path)?.readText()
